Clarification on the use of SQLFORM.grid.field_id
Text added at L 2500: "This is useful when the grid query is a join of several tables. Any action button on the grid(add record, view, edit, delete) will work over db.mytable." 

I think that the use of file_id requires clarification because I stumbled onto an issue when trying to make a grid with several tables joined and had no idea about how to control over which table where operating the action buttons of the grid.