In belongs section, example references wrong table
In the 'belongs' section, under the example for "those cases where a nested select is required and the look-up field is a reference..." the example defines table person, then table thing, with owner_id referring to itself (thing).  In this example, owner_id should reference the person table, not the thing table.